pyproject.toml
[project]
name = "wfcrai-agent-utils"
version = "0.1.0"
description = "Agent utilities and connectors package"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "openai>=1.65.0",
    "atlassian-python-api>=4.0.0",
    "dotenv>=0.9.9",
    "google-cloud-bigquery>=3.31.0",
    "langfuse>=2.60.2",
    "openai>=1.65.5",
    "toml>=0.10.2",
    "google-cloud-storage>=3.1.0",
]
